Move CANBoard class to CANMotor and remove CANBoard classes from real_world and FakeCANBoard.
